# docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_api_handler.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  robot_scheduler_internal_model_dto.LoginRequest:
    properties:
      password:
        description: 密码
        type: string
      userName:
        description: 用户名
        type: string
    required:
    - password
    - userName
    type: object
...
  robot_scheduler_internal_model_dto.TaskCreateRequest:
    properties:
      extraInfo:
        description: 扩展信息
        type: string
      semanticMapId:
        description: 对应的语义地图id
        type: integer
      taskInfo:
        description: 任务信息
        type: string
      userName:
        description: 编辑人员
        type: string
    required:
    - semanticMapId
    - taskInfo
    - userName
    type: object
  robot_scheduler_internal_model_dto.TaskUpdateRequest:
    properties:
      extraInfo:
        description: 扩展信息
        type: string
      semanticMapId:
        description: 对应的语义地图id
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/robot_scheduler_internal_model_entity.TaskStatus'
        description: 任务状态
      taskInfo:
        description: 任务信息
        type: string
      userName:
        description: 编辑人员
        type: string
    type: object
  robot_scheduler_internal_model_dto.UserCreateRequest:
    properties:
      extraInfo:
        description: 扩展信息
        type: string
      password:
        description: 密码
        minLength: 6
        type: string
      role:
        allOf:
        - $ref: '#/definitions/robot_scheduler_internal_model_entity.RoleType'
        description: 角色
        enum:
        - administrator
        - manager
        - operator
        - user
      userName:
        description: 用户名
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - role
    - userName
    type: object
  robot_scheduler_internal_model_dto.UserUpdateRequest:
    properties:
      extraInfo:
        description: 扩展信息
        type: string
      isLocked:
        description: 是否锁定
        enum:
        - 0
        - 1
        type: integer
      password:
        description: 密码
        type: string
      role:
        allOf:
        - $ref: '#/definitions/robot_scheduler_internal_model_entity.RoleType'
        description: 角色
    type: object
...
  robot_scheduler_internal_model_entity.RoleType:
    enum:
    - administrator
    - manager
    - operator
    - user
    type: string
    x-enum-comments:
      RoleAdministrator: 超级管理员
      RoleManager: 管理员
      RoleOperator: 操作员
      RoleUser: 普通用户
    x-enum-descriptions:
    - 超级管理员
    - 管理员
    - 操作员
    - 普通用户
    x-enum-varnames:
    - RoleAdministrator
    - RoleManager
    - RoleOperator
    - RoleUser
  robot_scheduler_internal_model_entity.TaskStatus:
    enum:
    - pending
    - running
    - completed
    - failed
    - cancelled
    type: string
    x-enum-comments:
      TaskStatusCancelled: 已取消
      TaskStatusCompleted: 已完成
      TaskStatusFailed: 失败
      TaskStatusPending: 待执行
      TaskStatusRunning: 执行中
    x-enum-descriptions:
    - 待执行
    - 执行中
    - 已完成
    - 失败
    - 已取消
    x-enum-varnames:
    - TaskStatusPending
    - TaskStatusRunning
    - TaskStatusCompleted
    - TaskStatusFailed
    - TaskStatusCancelled
host: localhost:8080
info:
  contact: {}
  description: 机器人调度系统后台管理接口
  termsOfService: http://swagger.io/terms/
  title: 机器人调度系统 API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 用户登录，返回JWT token
      parameters:
      - description: 登录信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/robot_scheduler_internal_model_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "401":
          description: 用户名或密码错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      summary: 用户登录
      tags:
      - 认证
  /auth/logout:
    post:
      consumes:
      - application/json
      description: 用户退出登录
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      security:
      - BearerAuth: []
      summary: 用户退出
      tags:
      - 认证
...
  /tasks:
    get:
      consumes:
      - application/json
      description: 查询所有任务
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      security:
      - BearerAuth: []
      summary: 查询任务列表
      tags:
      - 任务管理
    post:
      consumes:
      - application/json
      description: 创建新任务
      parameters:
      - description: 任务信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/robot_scheduler_internal_model_dto.TaskCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      security:
      - BearerAuth: []
      summary: 创建任务
      tags:
      - 任务管理
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      description: 删除任务（软删除）
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      security:
      - BearerAuth: []
      summary: 删除任务
      tags:
      - 任务管理
    get:
      consumes:
      - application/json
      description: 根据ID获取任务信息
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      security:
      - BearerAuth: []
      summary: 获取任务
      tags:
      - 任务管理
    put:
      consumes:
      - application/json
      description: 更新任务信息
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/robot_scheduler_internal_model_dto.TaskUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      security:
      - BearerAuth: []
      summary: 更新任务
      tags:
      - 任务管理
  /users:
    get:
      consumes:
      - application/json
      description: 查询所有用户
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      summary: 查询用户列表
      tags:
      - 用户管理
    post:
      consumes:
      - application/json
      description: 创建新用户
      parameters:
      - description: 用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/robot_scheduler_internal_model_dto.UserCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      summary: 创建用户
      tags:
      - 用户管理
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: 删除用户（软删除）
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      summary: 删除用户
      tags:
      - 用户管理
    get:
      consumes:
      - application/json
      description: 根据ID获取用户信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      summary: 获取用户
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      description: 更新用户信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/robot_scheduler_internal_model_dto.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      summary: 更新用户
      tags:
      - 用户管理
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"